Rename dashboard and bots pages
