+++ bots.html
@@ -41,6 +41,5 @@
     </div>
   </div>
   <script src="/js/main.js" type="module"></script>
-  <script src="/js/bots.js" type="module"></script>
 </body>
 </html>
