+++ dashboard.html
@@ -55,7 +55,6 @@
     </div>
   </div>
   <script src="/js/main.js" type="module"></script>
-  <script src="/js/dashboard.js" type="module"></script>
   <script src="/js/notifications.js" type="module"></script>
 </body>
 </html>
