+++ sw.js
@@ -6,8 +6,8 @@ const urlsToCache = [
   '/manifest.json',
   '/js/main.js',
   '/js/auth.js',
-  '/js/bots.js',
-  '/js/dashboard.js',
+  '/bots.html',
+  '/dashboard.html',
   '/js/inspiration.js',
   '/js/builder.js',
   '/js/workshop.js',
